refactor: exception 패키지 추가
