+++ src/main/java/com/example/moviewbackend/exception/CustomResponseException.java
@@ -1,4 +1,4 @@
-package com.example.moviewbackend;
+package com.example.moviewbackend.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
